src/main/java/org/avlasov/leetcodeproblems/easy/ImplementStrStr.java
package org.avlasov.leetcodeproblems.easy;

/**
 * Created By artemvlasov on 26/06/2018
 *
 * Implement strStr().
 *
 * Return the index of the first occurrence of needle in haystack, or -1 if needle is not part of haystack.
 *
 * Example 1:
 *
 * Input: haystack = "hello", needle = "ll"
 * Output: 2
 *
 * Example 2:
 *
 * Input: haystack = "aaaaa", needle = "bba"
 * Output: -1
 *
 * Clarification:
 *
 * What should we return when needle is an empty string? This is a great question to ask during an interview.
 *
 * For the purpose of this problem, we will return 0 when needle is an empty string. This is consistent to C's strstr() and Java's indexOf().
 **/
public class ImplementStrStr {

    public int strStr(String haystack, String needle) {
        if (haystack == null || needle == null || (haystack.length() < needle.length())) return -1;
        if (needle.isEmpty()) return 0;
        int haystackLength = haystack.length();
        int needleLength = needle.length();
        for (int i = 0; i <= haystackLength - needleLength; i++) {
            if (haystack.charAt(i) == needle.charAt(0) && haystack.substring(i, i + needleLength).equals(needle)) {
                return i;
            }
        }
        return -1;
    }

}
